Additional User Profiling Options!
Added more options for the user to choose from. More specifically, food choices and activity choices. These will be used to determine the target value for the TFIDF models, which are all implemented into our backend! We still need to implement functionality to the buttons as well as displaying all of the information.