Updated FetchSelectedAddresses File
Currently, this section isn’t working because the TFIDF doesn’t properly generate options from the correct city. In this FetchSelectedAddresses it requires a restaurant name, but it also asks for a city. The issue is that the restaurant may have the same name but a different matching city. This will work once the models are refined.